RpcLayer: Replace a Warning with Debug on MessageHeader check failed
There is a number of different reasons for 'invalid' Header.
We have explicit verbose warning messages in important cases but there
is no reason to warn about casual and sometimes expected fails.

For example, is a client has no up-to-date server salt, it can send
a message with any (outdated or zero) salt to make the server correct
it and send back the actual server salt. This commit hide warning in
such cases.